requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
coloredlogs==15.0.1
docopt==0.6.2
flatbuffers==23.5.26
humanfriendly==10.0
idna==3.6
mpmath==1.3.0
num2words==0.5.13
numpy==1.26.2
onnxruntime==1.16.3
packaging==23.2
piper-phonemize==1.1.0
piper-tts==1.2.0
protobuf==4.25.1
requests==2.31.0
sympy==1.12
urllib3==2.1.0
